Avoiding timed-attacks in Security::checkHash

+++ ext/kernel/operators.c
@@ -758,33 +758,33 @@ double zephir_safe_div_long_long(long op1, long op2 TSRMLS_DC) {
 		zend_error(E_WARNING, "Division by zero");
 		return 0;
 	}
-	return op1 / op2;
+	return (double) op1 / (double) op2;
 }
 
 /**
- * Do safe divisions between two longs
+ * Do safe divisions between two long/double
  */
 double zephir_safe_div_long_double(long op1, double op2 TSRMLS_DC) {
 	if (!op2) {
 		zend_error(E_WARNING, "Division by zero");
 		return 0;
 	}
-	return op1 / op2;
+	return (double) op1 / op2;
 }
 
 /**
- * Do safe divisions between two longs
+ * Do safe divisions between two double/long
  */
 double zephir_safe_div_double_long(double op1, long op2 TSRMLS_DC) {
 	if (!op2) {
 		zend_error(E_WARNING, "Division by zero");
 		return 0;
 	}
-	return op1 / op2;
+	return op1 / (double) op2;
 }
 
 /**
- * Do safe divisions between two longs
+ * Do safe divisions between two doubles
  */
 double zephir_safe_div_double_double(double op1, double op2 TSRMLS_DC) {
 	if (!op2) {
@@ -794,34 +794,46 @@ double zephir_safe_div_double_double(double op1, double op2 TSRMLS_DC) {
 	return op1 / op2;
 }
 
+/**
+ * Do safe divisions between two zval/long
+ */
 double zephir_safe_div_zval_long(zval *op1, long op2 TSRMLS_DC) {
 	if (!op2) {
 		zend_error(E_WARNING, "Division by zero");
 		return 0;
 	}
-	return zephir_get_numberval(op1) / op2;
+	return ((double) zephir_get_numberval(op1)) / (double) op2;
 }
 
+/**
+ * Do safe divisions between two zval/double
+ */
 double zephir_safe_div_zval_double(zval *op1, double op2 TSRMLS_DC) {
 	if (!op2) {
 		zend_error(E_WARNING, "Division by zero");
 		return 0;
 	}
-	return zephir_get_numberval(op1) / op2;
+	return ((double) zephir_get_numberval(op1)) / op2;
 }
 
+/**
+ * Do safe divisions between two long/zval
+ */
 double zephir_safe_div_long_zval(long op1, zval *op2 TSRMLS_DC) {
 	if (!zephir_get_numberval(op2)) {
 		zend_error(E_WARNING, "Division by zero");
 		return 0;
 	}
-	return op1 / zephir_get_numberval(op2);
+	return (double) op1 / ((double) zephir_get_numberval(op2));
 }
 
+/**
+ * Do safe divisions between two double/zval
+ */
 double zephir_safe_div_double_zval(double op1, zval *op2 TSRMLS_DC) {
 	if (!zephir_get_numberval(op2)) {
 		zend_error(E_WARNING, "Division by zero");
 		return 0;
 	}
-	return op1 / zephir_get_numberval(op2);
+	return op1 / ((double) zephir_get_numberval(op2));
 }
